Доработка 11 модуля
